minor refactoring, enhanced CreateTable method, moved migration into separate file
